Add requirements.txt

+++ init.sh
@@ -2,5 +2,14 @@
 
 python3 -m venv env
 source env/bin/activate
-pip install -r requirements.txt
+
+if [ ! -f "requirements.txt" ]; then
+  echo "Installing requirements and creating requirements.txt"
+  echo
+  pip install flask
+  pip freeze > requirements.txt
+else
+  pip install -r requirements.txt
+fi
+
 pycharm-community .
